apps/api/requirements.txt
fastapi==0.115.9
uvicorn[standard]==0.34.2
sqlalchemy==2.0.40
alembic==1.15.2
pydantic-settings==2.9.1
python-dotenv==1.1.0
python-multipart==0.0.20
chromadb==1.0.7
sentence-transformers==4.1.0
openai==1.76.0
pypdf==5.4.0
python-docx==1.1.2
pytest==8.3.5
